new button styling, leave guild functionality is now inside channel selector, selected channel now changes when changing guilds
