+++ frontend/src/app/App.tsx
@@ -8,18 +8,10 @@ import { useUserContext } from "./context/UserContext"
 export default function App() {
 	const { isAuthenticated, currentUser, sendWebSocketMessage } = useUserContext()
 
-	const createNewGuild = (guildName: string) => {
-		const guild = new CreateGuildEvent(currentUser.id, guildName)
-		sendWebSocketMessage(guild)
-	}
-
 	return (
 		<div>
 			{isAuthenticated ? (
-				<GuildsPanel
-					currentUser={currentUser}
-					createNewGuild={createNewGuild}
-				></GuildsPanel>
+				<GuildsPanel currentUser={currentUser}></GuildsPanel>
 			) : (
 				<RegisterLoginOption />
 			)}
