+++ frontend/src/app/components/Channel/Text/ChatPanel.tsx
@@ -28,10 +28,7 @@ export default function ChatPanel(props: ChatPanelProps) {
 	}
 
 	return (
-		<div
-			className="chat-panel flex flex-col h-full px-3 pb-5"
-			style={{ border: "2px solid blue" }}
-		>
+		<div className="chat-panel flex flex-col h-full px-3 pb-5">
 			<div
 				ref={chatContainerRef}
 				className="message-container flex flex-col gap-10 overflow-y-auto h-full pt-5 px-3 pb-3"
