+++ frontend/src/app/components/Guild/GuildSelector.tsx
@@ -9,49 +9,15 @@ import { UserDTO } from "@/app/DTOs/UserDTO"
 export default function GuildSelector(props: {
 	currentUser: UserDTO
 	selectGuild: (guild: GuildDTO) => void
-	createNewGuild: (guildName: string) => void
-	leaveGuild: (guildId: string) => void
 }) {
-	const [modalOpen, setModalOpen] = useState(false)
-
-	const openModal = () => {
-		setModalOpen(true)
-	}
-
-	const closeModal = () => {
-		setModalOpen(false)
-	}
-
 	return (
-		<aside
-			className="basis-[15%] py-5 px-4 border-r border-gray-700 flex flex-col"
-			style={{ border: "2px solid green" }}
-		>
+		<aside className="basis-[15%] py-5 border-r border-gray-800 flex flex-col">
 			<UserInfo currentUser={props.currentUser}></UserInfo>
-			<Divider className="my-2"></Divider>
-			<div>
-				<div className="my-3">
-					{props.currentUser.getGuilds().map((guild, index) => (
-						<GuildBtn
-							guild={guild}
-							selectGuild={props.selectGuild}
-							leaveGuild={props.leaveGuild}
-							key={index}
-						></GuildBtn>
-					))}
-					<Button
-						isIconOnly
-						className="flex justify-center items-center w-full my-2"
-						onClick={openModal}
-					>
-						<span className="material-symbols-outlined">add</span>
-					</Button>
-					<CreateGuildModal
-						isOpen={modalOpen}
-						onOpenChange={closeModal}
-						createNewGuild={props.createNewGuild}
-					/>
-				</div>
+			<Divider className="mb-2"></Divider>
+			<div className="my-3 flex flex-col gap-3 h-full">
+				{props.currentUser.getGuilds().map((guild, index) => (
+					<GuildBtn guild={guild} selectGuild={props.selectGuild} key={index}></GuildBtn>
+				))}
 			</div>
 		</aside>
 	)
