+++ frontend/src/app/components/Guild/MembersPanel.tsx
@@ -5,10 +5,7 @@ import { SenderDTO, UserDTO } from "@/app/DTOs/UserDTO"
 
 export default function MembersPanel(props: { members: SenderDTO[] }) {
 	return (
-		<div
-			className="basis-52 grow-0 shrink-0 border-l border-gray-700 flex flex-col justify-start items-start py-5 px-4 gap-5 overflow-y-auto"
-			style={{ border: "2px solid white" }}
-		>
+		<div className="basis-52 grow-0 shrink-0 border-l border-gray-800 flex flex-col justify-start items-start py-5 px-4 gap-5 overflow-y-auto">
 			{props.members.map((member, index) => (
 				<User
 					key={index}
